processing/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
deepface==0.0.93
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.66.2
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
Jinja2==3.1.4
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.15
ml-dtypes==0.4.1
mtcnn==1.0.0
namex==0.0.8
nats-py==2.9.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.5
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
retina-face==0.0.17
rich==13.9.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sounddevice==0.5.0
soupsieve==2.6
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
tf_keras==2.17.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
